Remove SYKEPENGER_MAX_DATE table and associated logic
Move test-related query function to test-scope
