+++ src/main/resources/db/migration/V32__Drop_sykepenger_max_data_table.sql
@@ -1,0 +1,1 @@
+DROP TABLE IF EXISTS SYKEPENGER_MAX_DATE;
